move superadmin CSP header setting from project-base to the framework package
