+++ packages/framework/src/Resources/views/Admin/Content/CspHeader/setting.html.twig
