+++ project-base/app/src/Controller/Admin/SideMenuConfigurationSubscriber.php
@@ -97,15 +97,6 @@ class SideMenuConfigurationSubscriber implements EventSubscriberInterface
         $closedDayMenu->addChild('closed_day_new', ['route' => 'admin_closedday_new', 'label' => t('New closed day'), 'display' => false]);
         $closedDayMenu->addChild('closed_day_edit', ['route' => 'admin_closedday_edit', 'label' => t('Holiday / internal day detail'), 'display' => false]);
 
-        $superadminSettingMenu = $settingsMenu->getChild('superadmin');
-
-        if ($superadminSettingMenu !== null) {
-            $superadminSettingMenu->addChild('cspHeader', ['route' => 'admin_cspheader_setting', 'label' => t('Content-Security-Policy header')]);
-
-            $settingsMenu->removeChild('superadmin');
-            $settingsMenu->addChild($superadminSettingMenu);
-        }
-
         $constantsMenu = $settingsMenu->addChild('constants', ['label' => t('Language constants')]);
         $constantsListMenu = $constantsMenu->addChild('constants_list', ['route' => 'admin_languageconstant_list', 'label' => t('List of language constants')]);
         $constantsListMenu->addChild('constants_edit', ['route' => 'admin_languageconstant_edit', 'label' => t('Language constant translation'), 'display' => false]);
